+++ services/121-service/src/programs/utils/overwrite-fsp-display-name.helper.ts
@@ -1,6 +1,8 @@
 import { FspConfigurationEnum } from '../../fsp/enum/fsp-name.enum';
 import { FinancialServiceProviderEntity } from '../../fsp/financial-service-provider.entity';
+import { RegistrationViewEntity } from '../../registration/registration-view.entity';
 import { ProgramFspConfigurationEntity } from '../fsp-configuration/program-fsp-configuration.entity';
+import { ProgramEntity } from '../program.entity';
 
 export function overwriteProgramFspDisplayName(
   programFinancialServiceProviders: FinancialServiceProviderEntity[],
@@ -31,4 +33,34 @@ export function overwriteProgramFspDisplayName(
   }
 
   return overwrittenProgramFinancialServiceProviders;
+}
+
+export function getFspDisplayNameMapping(
+  program: ProgramEntity,
+): Record<string, JSON> {
+  if (!program.financialServiceProviders || !program.programFspConfiguration) {
+    throw new Error(
+      `getFspDisplayNameMapping: Should be used with a program entity relations ['financialServiceProviders', 'programFspConfiguration']`,
+    );
+  }
+  const map = {};
+  if (program.financialServiceProviders.length > 0) {
+    program.financialServiceProviders = overwriteProgramFspDisplayName(
+      program.financialServiceProviders,
+      program.programFspConfiguration,
+    );
+  }
+  for (const fsp of program.financialServiceProviders) {
+    map[fsp.fsp] = fsp.displayName;
+  }
+  return map;
+}
+
+export function overwriteFspDisplayName(
+  registration: RegistrationViewEntity,
+  fspDisplayNameMapping: Record<string, JSON>,
+): JSON {
+  if (registration.financialServiceProvider) {
+    return fspDisplayNameMapping[registration.financialServiceProvider];
+  }
 }