+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.116.11...main)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.116.13...main)
 
 ### Removed
 
@@ -13,6 +13,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AW-App interface code and all related references
 
 ---
+
+## [1.116.13](https://github.com/global-121/121-platform/compare/v1.116.12...v1.116.13)- 2024-04-29
+
+### Fixed
+
+- FSP in Duplicates allow empty FSP
+
+### Added
+
+- scope to duplicate export
+
+## [1.116.12](https://github.com/global-121/121-platform/compare/v1.116.11...v1.116.12)- 2024-04-29
+
+### Fixed
+
+- FSP in Duplicates export
 
 ## [1.116.11](https://github.com/global-121/121-platform/compare/v1.116.10...v1.116.11)- 2024-04-29